feat(model): free previously unallocated resources and optimize training loop

+++ Project/mlp_codes/model.h
@@ -2,18 +2,28 @@
 #define MODEL_H
 
 // ----------------------
-// Variables globales
+// Global variables
 // ----------------------
 extern int num_examples;
 extern int nn_input_dim;
 extern int nn_output_dim;
-extern double reg_lambda;
-extern double epsilon;
-extern double *X;
+
+extern float reg_lambda;
+extern float epsilon;
+
+// Dataset
+extern float *X;
 extern int *y;
 
-// Fonctions principales
-double calculate_loss(double *W1, double *b1, double *W2, double *b2, int nn_hdim);
-void build_model(int nn_hdim, int num_passes, int print_loss);
+// ----------------------
+// Functions
+// ----------------------
+float calculate_loss(float *W1, float *b1,
+                     float *W2, float *b2,
+                     int nn_hdim);
+
+void build_model(int nn_hdim,
+                 int num_passes,
+                 int print_loss);
 
 #endif
